requirements.txt
requests
Pillow
cairosvg
